requirements.txt
fastapi
uvicorn
httpx
numpy
pyyaml
